+++ docs/html/search/files_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['compat_2eh_82',['compat.h',['../compat_8h.html',1,'']]],
-  ['coroutine_2eh_83',['Coroutine.h',['../Coroutine_8h.html',1,'']]]
+  ['compat_2eh_83',['compat.h',['../compat_8h.html',1,'']]],
+  ['coroutine_2eh_84',['Coroutine.h',['../Coroutine_8h.html',1,'']]]
 ];