+++ docs/html/search/functions_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['getjump_89',['getJump',['../classace__routine_1_1CoroutineTemplate.html#a84d4aaee5d3ba3750e78574908f2ba11',1,'ace_routine::CoroutineTemplate']]],
-  ['getstatus_90',['getStatus',['../classace__routine_1_1CoroutineTemplate.html#ae849b9399fefd59231c31c96c3ee19cf',1,'ace_routine::CoroutineTemplate']]]
+  ['getjump_90',['getJump',['../classace__routine_1_1CoroutineTemplate.html#a84d4aaee5d3ba3750e78574908f2ba11',1,'ace_routine::CoroutineTemplate']]],
+  ['getstatus_91',['getStatus',['../classace__routine_1_1CoroutineTemplate.html#ae849b9399fefd59231c31c96c3ee19cf',1,'ace_routine::CoroutineTemplate']]]
 ];