+++ docs/html/search/functions_5.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['read_105',['read',['../classace__routine_1_1Channel.html#a931ce2344667a4b36e76bb4f8f3b70a0',1,'ace_routine::Channel']]],
-  ['reset_106',['reset',['../classace__routine_1_1CoroutineTemplate.html#aac6c42a690d4f64942c7603b9890de9b',1,'ace_routine::CoroutineTemplate']]],
-  ['resume_107',['resume',['../classace__routine_1_1CoroutineTemplate.html#aab8a1eba56fb3d10273ce3c62263718b',1,'ace_routine::CoroutineTemplate']]],
-  ['runcoroutine_108',['runCoroutine',['../classace__routine_1_1CoroutineTemplate.html#a3da6aa3220b54203f050dc97d51b7070',1,'ace_routine::CoroutineTemplate']]]
+  ['read_106',['read',['../classace__routine_1_1Channel.html#a931ce2344667a4b36e76bb4f8f3b70a0',1,'ace_routine::Channel']]],
+  ['reset_107',['reset',['../classace__routine_1_1CoroutineTemplate.html#aac6c42a690d4f64942c7603b9890de9b',1,'ace_routine::CoroutineTemplate']]],
+  ['resume_108',['resume',['../classace__routine_1_1CoroutineTemplate.html#aab8a1eba56fb3d10273ce3c62263718b',1,'ace_routine::CoroutineTemplate']]],
+  ['runcoroutine_109',['runCoroutine',['../classace__routine_1_1CoroutineTemplate.html#a3da6aa3220b54203f050dc97d51b7070',1,'ace_routine::CoroutineTemplate']]]
 ];