+++ docs/html/search/typedefs_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['coroutine_137',['Coroutine',['../Coroutine_8h.html#a9971a0ed5c7ba49921d0aa2dbc8f83b0',1,'ace_routine']]]
+  ['coroutine_139',['Coroutine',['../Coroutine_8h.html#a9971a0ed5c7ba49921d0aa2dbc8f83b0',1,'ace_routine']]]
 ];